rename router and request in protocol package to op and service router
